harfeast_env/pyproject.toml
[project]
name = "harfeast-env"
version = "0.1.0"
description = "HarFeast management consulting RL environment - OpenEnv compatible"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["harfeast_env", "harfeast_openenv"]
